feat: share debug logger
