+++ examples/oft-solana/tasks/common/sendOFT.ts
@@ -7,7 +7,8 @@ import { EvmArgs, sendEvm } from '../evm/sendEvm'
 import { SolanaArgs, sendSolana } from '../solana/sendSolana'
 
 import { SendResult } from './types'
-import { DebugLogger, KnownOutputs, KnownWarnings, getBlockExplorerLink } from './utils'
+import { DebugLogger, KnownOutputs, KnownWarnings } from '@layerzerolabs/io-devtools'
+import { getBlockExplorerLink } from './utils'
 
 interface MasterArgs {
     srcEid: number
