+++ examples/oft-solana/tasks/common/wire.ts
@@ -19,13 +19,12 @@ import { findSolanaEndpointIdInGraph } from '../solana/utils'
 
 import { publicKey as publicKeyType } from './types'
 import {
-    DebugLogger,
-    KnownErrors,
     createSdkFactory,
     createSolanaConnectionFactory,
     createSolanaSignerFactory,
     getSolanaUlnConfigPDAs,
 } from './utils'
+import { DebugLogger, KnownErrors } from '@layerzerolabs/io-devtools'
 
 import type { SignAndSendTaskArgs } from '@layerzerolabs/devtools-evm-hardhat/tasks'
