+++ examples/oft-solana/tasks/solana/sendSolana.ts
@@ -12,7 +12,7 @@ import { addressToBytes32 } from '@layerzerolabs/lz-v2-utilities'
 import { oft } from '@layerzerolabs/oft-v2-solana-sdk'
 
 import { SendResult } from '../common/types'
-import { DebugLogger, KnownErrors } from '../common/utils'
+import { DebugLogger, KnownErrors } from '@layerzerolabs/io-devtools'
 
 import { parseDecimalToUnits, silenceSolana429 } from './utils'
