+++ packages/io-devtools/src/stdio/index.ts
@@ -1,3 +1,4 @@
 export * from './logger'
 export * from './printer'
 export * from './prompts'
+export * from './debugLogger'
